Fix PipesModule imports broken after rename.
